[Update] Update src/utils/image-processor.js and corresponding codes
