+++ src/components/TopNews.js
@@ -1,7 +1,7 @@
 'use strict'
 import { INTERACTIVE_ARTICLE_STYLE } from '../constants/index'
 import { date2yyyymmdd } from '../lib/date-transformer'
-import { getImageSrc, getImageSrcSet } from '../utils/index'
+import { getArticleImageSrc, getArticleImageSrcSet } from '../utils/index'
 import { Link } from 'react-router'
 import Hexagon from './Hexagon'
 import Slider from 'react-flex-carousel'
@@ -30,8 +30,8 @@ export default class TopNews extends Component {
           const pubDate = date2yyyymmdd(a.publishedDate, '.')
           let cats = get(a, 'categories', [])
           let catDisplay = get(cats, [ 0, 'name' ], '專題')
-          let imageSet = getImageSrcSet(a)
-          let image = getImageSrc(a)
+          let imageSet = getArticleImageSrcSet(a)
+          let image = getArticleImageSrc(a)
           return (
             <Link key={a.id} to={'/a/' + a.slug} target={a.style === INTERACTIVE_ARTICLE_STYLE ? '_self' : undefined}>
               <img src={image} alt={a.slug} srcSet={imageSet} />
