+++ src/utils/index.js
@@ -1,10 +1,10 @@
-export * from './article-image-processor'
 export * from './formaturl'
 export * from './denormalize-articles'
 export * from './detect-device'
 export * from './get-embedded-query-param'
 export * from './get-list-id'
 export * from './get-screen-type'
+export * from './image-processor'
 export * from './replace-storage-url-prefix'
 export * from './url-transformer'
 export * from './check-page-type'
